changing returning Transaction id from responses to Gameserver's

+++ Platipus.Wallet.Api/Application/Services/EvenbetGamesApi/External/EvenbetLoginRequest.cs
@@ -46,7 +46,7 @@ public sealed record EvenbetLoginRequest
 
             var response = new EvenbetLoginResponse(
                 loginData.Token,
-                (int)loginData.Balance,
+                loginData.Balance,
                 loginData.Currency,
                 loginData.Nickname,
                 DateTimeOffset.Now.ToUnixTimeSeconds().ToString(),
